XmlHttpRequest: API change for perform functions.
The proposed change is to offer more possibilities for the body
contents of a request. While only form-like bodies were available,
this commit introduces two more possibilities: an arbitrary string and
a blob object (for file upload notably).

For this the signature of perform functions was changed: arguments
`post_args` and `form_args` were replaced by a `contents` argument,
which makes the behaviour of the function a bit more readable IMHO.

In addition to breaking the API, there's one regression, namely that
it is not possible to combine POST arguments with a FormData object,
though I'm not sure if that's critical.