Port all PPX's to ppxlib
Before, the two extension node rewriters in the compiler and in ppx_js
were based on the omp driver and the omp Ast_mapper. The deriver in
ppx_deriving_json was already partly based on ppxlib, but also on
omp's Ast_408. All of those omp modules will be dropped in omp 2.0.0.

That's why this commit ports those 3 PPX's to ppxlib.