Optimize all tail calls when effects are enabled
With the CPS transform used to support effect handlers, most function
calls will be in tail position. This change allows to generate
intermediate functions that check that the stack does not become too
deep before calling an OCaml function, like we have caml_callN to call
functions that are expected to have N parameters.