Effects: omit some stack depth checks
We omit stack checks when jumping from one block to another within a
function, except for backward edges. Stack checks are also omitted when
calling the function continuations.

We have to check the stack depth in `caml_alloc_stack` for the test
`evenodd.ml` to succeed. Otherwise, popping all the fibers exhaust the
JavaScript stack. We don't have this issue with the OCaml runtime since
it allocates one stack per fiber.