Run compiler against stdlib_modern.js
This commit has the unfortunate side-effect of printing twice any compiler
warnings produced during a call to compile_and_run. Since there happens
to be only one such instance, this seems acceptable for now.

Running the compiler twice means that the tests have to be approximately
idempotent. This commit fixes the two tests where this is not the case.