Don't hold on to channels after they've been closed
In my use case, this is responsible for O(100MiB) of
memory usage, because we hold on to multiple copies of time zone data, and
each file has a 65KiB buffer that doesn't get released.

The file io.js provides the same runtime as the C runtime from [2]. There's a global
array of all existing channels, and this never gets released even if a channel is closed.
In the C runtime, this does eventually happen in a call to [caml_finalize_channel] which
is called by the GC, but I don't think we have that ability in js_of_ocaml.

This feature makes [chanid] an object rather than an integer, so we can let
the javascript garbage collector take care of it.

The object [caml_ml_channels] continues to exist, but its only remaining
function is to provide "override" functionality in the way that ppx_expect
needs it.

Some useful references:

[1]: https://v2.ocaml.org/releases/5.1/api/Stdlib.html#2_Outputfunctionsonstandardoutput
[2]: https://github.com/ocaml/ocaml/blob/trunk/runtime/io.c#L163
[3]: https://github.com/janestreet/ppx_expect/blob/master/runtime/runtime.js