Implement a stack of debug information
The goal is to avoid source location information from leaking to
unrelated pieces of code. For instance, if we have some debugging
information missing after an if statement or within an else branch we
should assume that we remain at the same location in the source code,
and therefore ignore any location information from a branch of the if.
This issue happens in the compilation of matches, both by the OCaml
compiler and Jsoo.