Improved statement locations
Statements will get the earliest location of the expressions performing
a call it contains, if there is any. Then, for an expression `if (e)
{...}`, we don't move first to the location of the conditional before
going back to the location within `e` which might be strictly before.

When queueing expressions, we ignore their locations if they don't
perform a call; when flushed, we use the current location. This again
prevent locations to be reordered.