Make sure there is a location at the start of functions
If the first block of the function starts with an event, we use the
event's location. This is useful in case of tail-calls, where some code
is generated before this first block. If we don't have an event, we set
the initial location to unkown to prevent previous locations to bleed
into the function body.