Output a Wasm binary module rather than a text file when possible
This is faster to write a binary file, and then for binaryen to parse a
binary file.

We still output a Wasm text file when source maps are enabled (we don't
yet supported writing a source map file together with a Wasm binary
module) or when `--debug wat` is set.