Runtime/wasm: fix method lookup, take 2
The method table contains the following. Index 0 contains the block's
tag (0). Index 1 contains the number of public methods. Index 2 contains
a mask. Then, we have each method followed by its tag. The table is
completed with dummy non-integer values to reach a power-of-two size
(not including the first three indices).